+++ include/royale_ros/camera_nodelet.h
@@ -33,6 +33,7 @@
 #include <royale_ros/Start.h>
 #include <royale_ros/Stop.h>
 #include <royale_ros/SetExposureTime.h>
+#include <royale_ros/SetExposureTimes.h>
 #include <royale.hpp>
 
 namespace royale_ros
@@ -66,6 +67,7 @@ namespace royale_ros
     // Handlers for topic subscriptions
     //
     void SetExposureTimeCb(const royale_ros::SetExposureTime::ConstPtr& msg);
+    void SetExposureTimesCb(const royale_ros::SetExposureTimes::ConstPtr& msg);
 
     //
     // Royale callback, basically functions as our main
@@ -111,6 +113,7 @@ namespace royale_ros
     ros::Timer timer_;
 
     ros::Subscriber exp_time_sub_;
+    ros::Subscriber exp_times_sub_;
 
     std::unique_ptr<image_transport::ImageTransport> it_;
     std::vector<ros::Publisher> cloud_pubs_;