+++ ChangeLog.md
@@ -5,6 +5,7 @@
 * Updated compiler flags to C++14
 * Unlocked (some, interesting) Level-2 access features (assuming a proper access key)
 * Removed (unnecessary) `config.launch`
+* Added SetExposureTimes topic (Level 2 access)
 
 ## This file has started tracking royale-ros at 0.1.0